requirements.txt
flask
flask_sqlalchemy
azure-cognitiveservices-vision-computervision
msrest
opencv-python
requests
pika
stomp.py
pymysql